+++ packages/types/src/extrinsic/signedExtensions/index.ts
@@ -12,16 +12,18 @@ import { substrate } from './substrate';
 // we define the extra fields not as a Tuple, but rather as a struct so they can be named. These will be expanded
 // into the various fields when added to the payload (we only support V4 onwards with these, V3 and earlier are
 // deemed fixed and non-changeable)
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const allExtensions: ExtDef = {
+export const allExtensions: ExtDef = {
   ...substrate,
-  ...shell,
   ...polkadot,
+  ...shell,
   ...statemint
 };
 
-// the v4 signed extensions (the order is important here, as applied by default)
-const defaultExtensions: Array<keyof typeof allExtensions> = [
+// the v4 signed extensions prior to the point of exposing these to the metadata.
+// This may not match 100% with the current defaults and are used when not specified
+// in the metadata (which is for very old chains). The order is important here, as
+// applied by default
+export const fallbackExtensions: Array<keyof typeof allExtensions> = [
   'CheckVersion',
   'CheckGenesis',
   'CheckEra',
@@ -31,18 +33,16 @@ const defaultExtensions: Array<keyof typeof allExtensions> = [
   'CheckBlockGasLimit'
 ];
 
-function findUnknownExtensions (extensions: string[], userExtensions: ExtDef = {}): string[] {
+export function findUnknownExtensions (extensions: string[], userExtensions: ExtDef = {}): string[] {
   const names = [...Object.keys(allExtensions), ...Object.keys(userExtensions)];
 
   return extensions.filter((key) => !names.includes(key));
 }
 
-function expandExtensionTypes (extensions: string[], type: keyof ExtInfo, userExtensions: ExtDef = {}): ExtTypes {
+export function expandExtensionTypes (extensions: string[], type: keyof ExtInfo, userExtensions: ExtDef = {}): ExtTypes {
   return extensions
     // Always allow user extensions first - these should provide overrides
     .map((key) => userExtensions[key] || allExtensions[key])
     .filter((info): info is ExtInfo => !!info)
     .reduce((result, info): ExtTypes => ({ ...result, ...info[type] }), {});
 }
-
-export { allExtensions, defaultExtensions, expandExtensionTypes, findUnknownExtensions };